Revert "fix: improved web push management (#3421)" (#4058)
