+++ overseerr-api.yml
@@ -3579,8 +3579,6 @@ paths:
                   type: string
                 p256dh:
                   type: string
-                userAgent:
-                  type: string
               required:
                 - endpoint
                 - auth
@@ -3588,88 +3586,6 @@ paths:
       responses:
         '204':
           description: Successfully registered push subscription
-  /user/{userId}/pushSubscriptions:
-    get:
-      summary: Get all web push notification settings for a user
-      description: |
-        Returns all web push notification settings for a user in a JSON object.
-      tags:
-        - users
-      parameters:
-        - in: path
-          name: userId
-          required: true
-          schema:
-            type: number
-      responses:
-        '200':
-          description: User web push notification settings in JSON
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  endpoint:
-                    type: string
-                  p256dh:
-                    type: string
-                  auth:
-                    type: string
-                  userAgent:
-                    type: string
-  /user/{userId}/pushSubscription/{key}:
-    get:
-      summary: Get web push notification settings for a user
-      description: |
-        Returns web push notification settings for a user in a JSON object.
-      tags:
-        - users
-      parameters:
-        - in: path
-          name: userId
-          required: true
-          schema:
-            type: number
-        - in: path
-          name: key
-          required: true
-          schema:
-            type: string
-      responses:
-        '200':
-          description: User web push notification settings in JSON
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  endpoint:
-                    type: string
-                  p256dh:
-                    type: string
-                  auth:
-                    type: string
-                  userAgent:
-                    type: string
-    delete:
-      summary: Delete user push subscription by key
-      description: Deletes the user push subscription with the provided key.
-      tags:
-        - users
-      parameters:
-        - in: path
-          name: userId
-          required: true
-          schema:
-            type: number
-        - in: path
-          name: key
-          required: true
-          schema:
-            type: string
-      responses:
-        '204':
-          description: Successfully removed user push subscription
   /user/{userId}:
     get:
       summary: Get user by ID
