+++ package.json
@@ -38,7 +38,6 @@
     "@supercharge/request-ip": "1.2.0",
     "@svgr/webpack": "6.5.1",
     "@tanem/react-nprogress": "5.0.30",
-    "@types/ua-parser-js": "^0.7.36",
     "ace-builds": "1.15.2",
     "axios": "1.3.4",
     "axios-rate-limit": "1.3.0",
@@ -91,7 +90,6 @@
     "swagger-ui-express": "4.6.2",
     "swr": "2.0.4",
     "typeorm": "0.3.12",
-    "ua-parser-js": "^1.0.35",
     "web-push": "3.5.0",
     "winston": "3.8.2",
     "winston-daily-rotate-file": "4.7.1",
