+++ server/entity/UserPushSubscription.ts
@@ -1,10 +1,4 @@
-import {
-  Column,
-  CreateDateColumn,
-  Entity,
-  ManyToOne,
-  PrimaryGeneratedColumn,
-} from 'typeorm';
+import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { User } from './User';
 
 @Entity()
@@ -24,14 +18,8 @@ export class UserPushSubscription {
   @Column()
   public p256dh: string;
 
-  @Column()
+  @Column({ unique: true })
   public auth: string;
-
-  @Column({ nullable: true })
-  public userAgent: string;
-
-  @CreateDateColumn({ nullable: true })
-  public createdAt: Date;
 
   constructor(init?: Partial<UserPushSubscription>) {
     Object.assign(this, init);
